requirements.txt
onnxruntime==1.13.1
opencv-python==4.6.0.66
Pillow==9.3.0
streamlit>=1.20.0